+++ src/main/java/com/blogApp/repositories/CommentRepo.java
@@ -3,5 +3,5 @@ package com.blogApp.repositories;
 import com.blogApp.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepo extends JpaRepository<Comment, Integer> {
+public interface CommentRepo extends JpaRepository<Comment, Long> {
 }
