+++ src/main/java/com/blogApp/repositories/PostRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface PostRepo extends JpaRepository<Post,Integer> {
+public interface PostRepo extends JpaRepository<Post,Long> {
 
     List<Post> findByUser(User user);
     List<Post> findByCategory(Category category);
