+++ src/main/java/com/blogApp/service/CommentService.java
@@ -3,6 +3,6 @@ package com.blogApp.service;
 import com.blogApp.Dto.CommentDto;
 
 public interface CommentService {
-    CommentDto createComment(CommentDto commentDto,Integer postId);
-    void deleteComment(Integer commentId );
+    CommentDto createComment(CommentDto commentDto,Long postId);
+    void deleteComment(Long commentId );
 }
