+++ src/main/java/com/blogApp/service/PostService.java
@@ -2,18 +2,21 @@ package com.blogApp.service;
 
 import com.blogApp.Dto.PostDto;
 import com.blogApp.Dto.PostResponse;
-import com.blogApp.entity.Post;
+
 
 import java.util.List;
 
 public interface PostService {
-    PostDto createPost(PostDto postDto,Integer userId,Integer categoryId);
-    PostDto updatePost(PostDto postDto,Integer postId);
-    void deletePost(Integer postId);
+    PostDto createPost(PostDto postDto,Long userId,Long categoryId);
+
+
+
+    PostDto updatePost(PostDto postDto, Long postId);
+    void deletePost(Long postId);
     PostResponse getAllPosts(Integer pageNumber , Integer pageSize,String sortBy);
-    PostDto getPostById(Integer postId);
-    List<PostDto> getPostsByCategory(Integer categoryId);
-     List<PostDto> getPostsByUser(Integer userId);
+    PostDto getPostById(Long postId);
+    List<PostDto> getPostsByCategory(Long categoryId);
+    List<PostDto> getPostsByUser(Long userId);
      List<PostDto> searchPosts(String keywords);
 
 }
